create simple setup script
