--- tools/publish.py
@@ -1,17 +1,0 @@
-#!/usr/bin/python
-#coding=utf-8
-
-import sys
-
-def _check_python_version():
-    major_ver = sys.version_info[0]
-    if major_ver > 2:
-        print ("The python version is %d.%d. But python 2.x is required. (Version 2.7 is well tested)\n"
-               "Download it here: https://www.python.org/" % (major_ver, sys.version_info[1]))
-        return False
-
-    return True
-
-if __name__ == '__main__':
-    if not _check_python_version():
-        exit()
